code refactor

+++ inc/Query.hpp
@@ -18,12 +18,13 @@ class Query
 	Query(void);
 	Query(Query const &copy);
 	bool ready;
+
   public:
 	int				   get_socket(void) const;
 	bool			   is_ready(void) const;
 	void			   form_request(void);
 	HTTPRequest const *get_request(void) const;
-	int				   recieve(void);
+	size_t			   recieve(void);
 	size_t			   send(std::string const &) const;
 	Query(struct pollfd *);
 	short getRevents(void) const;
